requirements.txt
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
dateparser==0.7.0
h5py==2.8.0
humanize==0.5.1
idna==2.5
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
newsapi-python==0.2.3
nltk==3.4
numpy==1.15.4
pendulum==2.0.4
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
regex==2018.11.22
requests==2.17.1
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snaptime==0.2.4
Theano==1.0.3
tzlocal==1.5.1
urllib3==1.21.1
